+++ chevah/compat/testing/testcase.py
@@ -95,6 +95,9 @@ class TwistedTestCase(TestCase):
     tests.
     """
 
+    # Number of second to wait for a deferred to have a result.
+    DEFERRED_TIMEOUT = 1
+
     # List of names for delayed calls which should not be considered as
     # required to wait for them when running the reactor.
     EXCEPTED_DELAYED_CALLS = []
@@ -206,7 +209,7 @@ class TwistedTestCase(TestCase):
             'Reactor took more than %.2f seconds to execute.' % timeout)
         self._reactor_timeout_failure = failure
 
-    def _initiateTestReactor(self, timeout=1):
+    def _initiateTestReactor(self, timeout):
         """
         Do the steps required to initiate a reactor for testing.
         """
@@ -340,7 +343,7 @@ class TwistedTestCase(TestCase):
                 raise_failure('delayed calls', delayed_str)
 
     def runDeferred(
-            self, deferred, timeout=1, debug=False, prevent_stop=False):
+            self, deferred, timeout=None, debug=False, prevent_stop=False):
         """
         Run the deferred in the reactor loop.
 
@@ -366,6 +369,9 @@ class TwistedTestCase(TestCase):
         """
         if not isinstance(deferred, Deferred):
             raise AssertionError('This is not a deferred.')
+
+        if timeout is None:
+            timeout = self.DEFERRED_TIMEOUT
 
         try:
             self._initiateTestReactor(timeout=timeout)
@@ -394,7 +400,7 @@ class TwistedTestCase(TestCase):
             self._runDeferred(result, timeout=timeout, debug=debug)
             result = deferred.result
 
-    def executeReactor(self, timeout=1, debug=False, run_once=False):
+    def executeReactor(self, timeout=None, debug=False, run_once=False):
         """
         Run reactor until no more delayed calls, readers or
         writers or threads are in the queues.
@@ -421,6 +427,9 @@ class TwistedTestCase(TestCase):
 
             self.assertStartsWith('211-Features:\n', result)
         """
+        if timeout is None:
+            timeout = self.DEFERRED_TIMEOUT
+
         self._initiateTestReactor(timeout=timeout)
 
         # Set it to True to enter the first loop.
@@ -470,7 +479,7 @@ class TwistedTestCase(TestCase):
                         '%s' % (self._reactorQueueToString()))
                 break
 
-            # Look at writters buffers:
+            # Look at writers buffers:
             if len(reactor.getWriters()) > 0:
                 have_callbacks = True
                 continue
@@ -504,7 +513,7 @@ class TwistedTestCase(TestCase):
         return raw_name.split(' ', 1)[0]
 
     def getDeferredFailure(
-            self, deferred, timeout=1, debug=False, prevent_stop=False):
+            self, deferred, timeout=None, debug=False, prevent_stop=False):
         """
         Run the deferred and return the failure.
 
@@ -652,7 +661,7 @@ class TwistedTestCase(TestCase):
                     deferred, result[0]))
 
     def getDeferredResult(
-            self, deferred, timeout=1, debug=False, prevent_stop=False):
+            self, deferred, timeout=None, debug=False, prevent_stop=False):
         """
         Run the deferred and return the result.